# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AetherGrid-Sovereign API",
    "description": "Operational API platform for the AetherGrid-Sovereign cascade prediction system.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scenarios/sync": {
      "post": {
        "tags": [
          "Scenarios & Predictions"
        ],
        "summary": "Run Scenario Sync",
        "description": "Runs a synchronous small-scenario inference and returns a safety certificate.",
        "operationId": "run_scenario_sync_api_v1_scenarios_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Run Scenario Sync Api V1 Scenarios Sync Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scenarios/async": {
      "post": {
        "tags": [
          "Scenarios & Predictions"
        ],
        "summary": "Submit Scenario Async",
        "description": "Submits a scenario to the asynchronous workers queue.",
        "operationId": "submit_scenario_async_api_v1_scenarios_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Submit Scenario Async Api V1 Scenarios Async Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scenarios/jobs/{job_id}": {
      "get": {
        "tags": [
          "Scenarios & Predictions"
        ],
        "summary": "Get Job Status",
        "description": "Retrieves status and logs/results of an asynchronous job.",
        "operationId": "get_job_status_api_v1_scenarios_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Status Api V1 Scenarios Jobs  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "Scenarios & Predictions"
        ],
        "summary": "Cancel Job",
        "description": "Cancels a queued or running job.",
        "operationId": "cancel_job_api_v1_scenarios_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Job Api V1 Scenarios Jobs  Job Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioPayload": {
        "properties": {
          "trigger_nodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trigger Nodes",
            "description": "Target node IDs to fail"
          },
          "hazard_intensity": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Hazard Intensity",
            "default": 0.8
          },
          "sensor_dropout": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Sensor Dropout",
            "default": 0.0
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "trigger_nodes"
        ],
        "title": "ScenarioPayload"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}